Stefan Seyfried <seife _at_ suse.de> writes;
i always believed "service" was a LSB thing, but could not find it in the 3.1
specification.
On SUSE systems, "service foo {start|stop|restart|status}" just calls the
init script, which has to support "status" anyway. I don't know about the
Pardus init scripts, but wouldn't it be, in the long term, better for you
if you just made your "service" behave like the "rest of the world"? :-)
Means: a patch to "service" to take "status" as an alias for "info" would
probably be as easy as patching this into pm-utils :-)

I might be wrong here, of course, and i am not LSB-enlightened at all.

So status added as an alias :)