files:
old = ["a", "b", "c"]
new = ["d", "e", "c"]

If all these files have same hash, we can use old[0] for all the
copy to new location operations.

So we have

a -> d
a -> e
a -> c

for relocations. 

But one problem is, old already has "c". 
So an "a -> c" sym link copy ends with "File exists" error.