cleanup rebuild-db
Delete most of the rebuild-db code. We only have filesdb as a conventional db.
Also this db file decreased in size from 120MB to 40MB by only holding
(path, pkg_name) as a key, value pair. Previous db code was holding the fileinfo
object, which can easily be taken from installdb.get_files(pkg) code if more
information is needed.

Filesdb is only for searching a path and pkg_name correlation. Unfortunately
this db is currently needed as some code in pisi - file conflict - needs
a fast has_file function. We could have not used this db and search linear
through the files.xml s like other package managers.

"The good thing about not having a db is not having a db" ...

So we currently need rebuild-db crap for recreating a filesdb. Rebuild-db
is bad because every bug of pisi is been tried to solved with this command.
If we can get rid of this, we can fix the real bugs.. but we are close..