speed up list_upgradable
"Premature optimization is the root of all evil."

But this pissed me of: pisi.api.list_upgradable was taking around 7.2 seconds.
Now it takes 0.5 seconds in my computer.

The very good thing with this new filesystem-piksemel db shows us that, there
is always a room for optimization. And this is good.