Repo order
We have cd, usb, remote and local. Any repo order is within these
medias. This is configured in pisi.conf. This resembles boot order.

For example there are two remote repos: contrib and pardus-2007

In pisi.conf we see that "usb" is before "remote". So when a usb
repository is found, it can be asked by kde and added to pisi repos.
Packages will be installed from usb repo now. If not found. fall back
happens.