Make a distinction between /usr/lib and /usr/libexec while checking for abandoned files.
Say that there exists a '/usr/libexec/hal_lpadmin' in the installDIR of
the package but '/usr/libexec' is not written to the Files section of
the pspec.xml. If there's another path '/usr/lib' written in the Files section,
the string method find() used in get_abandoned_files() matches '/usr/libexec'
because they have exactly the same beginning.

This can either be fixed by postfixing all the directory paths in pspec.xml's with
'/' or appending a '/' during the abandoned files checking procedure.

Thanks to Gokmen Gorgen for reporting this erratic behaviour.