pisi.api.add_repo and pisi.api.update_repo should not remove any repository by force. These are two seperate calls that just adds a repository and just updates a repository, respectively.
Checking distribution or architecture of a repository should be done by the pisi clients.
RepoDB has two methods for this (check_architecture, check_distribution) that return False when failed.
And any action desired should be taken after by the clients themselves. Removing a repository after a pisi
up is insane after a repository is added anyhow.

So for pisi-cli, from now on pisi add-repo will add, update and check the repository by default without asking
if you want to update the repository or not. If fetching the repository index from the given url fails, 
pisi cli will remove the repository as before.

Two options added to pisi add-repo: --ignore-check and --no-fetch

If --ignore-check is given, pisi ignores the arch/distro check after a _successful_ add and update phase.

--no-fetch is for not fetching the repository index. If this option is given, pisi will be unable to check
the arch and distro of the repository (because we have no index to check) and will warn about this to continue.

Summary: 
If a repository is added to pisi database, this means either it is checked or forced. Either way pisi will
use these repositories.