build: Change filename format of packages
The new format is:

    <name>-<version>-<release>-<distribution id>-<architecture>.pisi

    e.g. pisi-2.3.2-180-p11-x86_64.pisi

We will not use build numbers any more. This commit removes some of the code
related to build numbers. New pisi will not deal with build numbers.
Before the stable release, all buildno-related code will be removed.

It is now possible to use digits after a dash character (e.g. polkit-qt-1).

Distribution id is a short string to represent the distribution release
(e.g. p11 for Pardus 2011).

Architecture strings cannot start with a digit. This is needed to
distinguish old and new packages.