actionsapi/texlivemodules: Add exception for None values of "patterns"
Every format file has some keywords that are passed to the appropriate 
application during the build stage. However some of thes keywords, such
as "patterns" may have Null values (do not exist) or have empty string.
Thus, we add here an exception for Null values of "patterns". If no 
"patterns" keyword exist, than we create a new one with the parameter '-'
This parameter is default and is adviced from the TexLive developers itself.