Changed the behaviour of updating grub. The function accepts an argument to change the behaviour. There are three scenarios:
auto:   This is a toggle argument. It will append/remove the line
        blacklist=foo to grub.conf. If available it removes it, if not
        it appends it.
vendor: This will append the blacklist=foo line to grub.conf
os:     This will remove the blacklist=foo line from grub.conf

Besides that an escape is added for graphic cards that are not in driversDB.
This will cause that nothing happens to the grub.conf file.