build: Add support for build types
With build types, it is possible to use the same actions.py
for re-building the package with different options. In actions.py
scripts, get.buildTYPE() function can be used to get the current
build type.

A BuildType tag must be specified inside the Package tag of the related
binary package.