build: Add a method to determine delta packages to be created
This will be used by buildfarm. It accepts release numbers to create
delta packages for stable distribution releases. If the release number
is not given, it searches the binary package dirs for old packages and
creates delta packages. max_count can be specified to limit numbers of
delta packages.

It is possible to override search paths via search_paths parameter.

release and max_count parameters cannot be used at the same time.

This commit also removes the max_delta_count option from config file
since it is useless and confusing.